requirements.txt
streamlit==1.51.0
langchain==1.0.7
langchain-google-genai==3.0.3
langchain-community==0.4.1
langchain-text-splitters==1.0.0
faiss-cpu==1.12.0
google-generativeai==0.8.5
pandas==2.2.3
PyPDF2==3.0.1
python-dotenv==1.0.1
